+++ finagle-mux/src/main/scala/com/twitter/finagle/mux/ServerProcessor.scala
@@ -1,10 +1,19 @@
 package com.twitter.finagle.mux
 
-import com.twitter.finagle.{Dtab, FailureFlags, Filter, Path, Service}
+import com.twitter.finagle.Dtab
+import com.twitter.finagle.FailureFlags
+import com.twitter.finagle.Filter
+import com.twitter.finagle.Path
+import com.twitter.finagle.Service
 import com.twitter.finagle.context.Contexts
-import com.twitter.finagle.mux.transport.{Message, MuxFailure}
+import com.twitter.finagle.mux.transport.Message
+import com.twitter.finagle.mux.transport.MuxFailure
+import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finagle.tracing.Trace
-import com.twitter.util.{Future, Return, Throw}
+import com.twitter.io.Buf
+import com.twitter.util.Future
+import com.twitter.util.Return
+import com.twitter.util.Throw
 
 /**
  * Processor handles request, dispatch, and ping messages. Request
@@ -14,16 +23,18 @@ import com.twitter.util.{Future, Return, Throw}
  * (This arrangement permits interpositioning other filters to modify ping
  * or dispatch behavior, e.g., for testing.)
  */
-private[finagle] object ServerProcessor extends Filter[Message, Message, Request, Response] {
+private[finagle] class ServerProcessor(statsReceiver: StatsReceiver)
+    extends Filter[Message, Message, Request, Response] {
   import Message._
 
   private[this] val AlwaysEmpty = { _: Throwable => MuxFailure.Empty }
+  private[this] val contextBytesStat = statsReceiver.stat("request_context_bytes")
 
   private[this] def dispatch(
     tdispatch: Message.Tdispatch,
     service: Service[Request, Response]
   ): Future[Message] = {
-
+    contextBytesStat.add(contextBytes(tdispatch.contexts))
     Contexts.broadcast.letUnmarshal(tdispatch.contexts) {
       if (tdispatch.dtab.nonEmpty)
         Dtab.local ++= tdispatch.dtab
@@ -76,5 +87,15 @@ private[finagle] object ServerProcessor extends Filter[Message, Message, Request
     case Message.Tping(Message.Tags.PingTag) => Message.PreEncoded.FutureRping
     case Message.Tping(tag) => Future.value(Message.Rping(tag))
     case m => Future.exception(new IllegalArgumentException(s"Cannot process message $m"))
+  }
+
+  private[this] def contextBytes(contexts: Seq[(Buf, Buf)]): Int = {
+    var bytes = 0
+    contexts.foreach {
+      case (key, value) =>
+        bytes += key.length
+        bytes += value.length
+    }
+    bytes
   }
 }